requirements.txt
Django==5.0.4
python-decouple==3.8
Pillow==12.2.0
qrcode==8.2
psycopg2-binary==2.9.11
celery==5.6.3
redis==7.4.0
django_celery_results==2.6.0
gunicorn==25.3.0
deepface==0.0.99
tf-keras==2.18.0
opencv-python-headless==4.11.0.86